discordbot_dev/requirements.txt
discord.py==2.4.0
supabase==2.5.0
python-dotenv==1.0.1

